Updating my final lessons.
I can't seem to get the last section of my assignment working.  I have
spoken with your TA and fellow class mates and I am still stuck.

I had spent a bunch of time on the project so this was sort of pushed
back.  I will continue to work on it if its possible.